Se agregó: Valida privilegios para ver en caballeriza

+++ pages/caballerizas.php
@@ -1,6 +1,11 @@
 <?php
   include($_SERVER['DOCUMENT_ROOT'] . 'hipodromo/includes/header.inc.php');
   
+  if ((json_decode($db->getPrivilegiosByRol($_SESSION['rol']['pkrol_id'],1))->action != "success") AND (json_decode($db->getPrivilegiosByRol($_SESSION['rol']['pkrol_id'],5))->action != "success")) {
+      echo '<meta http-equiv="refresh" content="0;url=' . $_SESSION["last_uri"] . '">';
+      die();
+  }
+  
   if ($_SERVER["REQUEST_METHOD"] == "POST") {
       if (!empty($_POST['pkcab_id'])) {
           $answer = json_decode($db->deleteCaballeriza($_POST['pkcab_id']));
